move RegionEditor to reusable
